seperated out db interfaces to obey ISP and moved DB models to database package (leave room tests still fail)
