+++ src/main/java/Main.java
@@ -11,8 +11,8 @@ import ui.*;
 public class Main {
     private static final String API_URL = "https://api.sheety.co/78ad1edb28469578058ca4c58c3f478b/larklink";
     public static void main(String[] args) {
-        UserDBGateway userDBAccess = new UserDBAccess(API_URL);
-        RoomDBGateway roomDBAccess = new RoomDBAccess(API_URL);
+        UserDBAccess userDBAccess = new UserDBAccess(API_URL);
+        RoomDBAccess roomDBAccess = new RoomDBAccess(API_URL);
 
         UserPresenter userPresenter = new UserPresenter();
         UserInteractor userInteractor = new UserInteractor(userPresenter, userDBAccess);
