+++ src/main/java/database/RoomDBModel.java
@@ -1,4 +1,4 @@
-package models;
+package database;
 
 import java.util.List;
