+++ src/main/java/database/UserDBModel.java
@@ -1,17 +1,24 @@
-package models;
+package database;
 
-public class UserModel {
+public class UserDBModel {
+    private final int userID;
     private final String username;
     private final String password;
 
-    public UserModel(String username, String password) {
+    public UserDBModel(int userID, String username, String password) {
+        this.userID = userID;
         this.username = username;
         this.password = password;
+    }
+
+    public int getUserID() {
+        return this.userID;
     }
 
     public String getUsername() {
         return this.username;
     }
+
     public String getPassword() {
         return this.password;
     }
