+++ src/main/java/host_room/HostRoomDBGateway.java
@@ -1,0 +1,10 @@
+package host_room;
+
+import database.RoomDBModel;
+
+import java.util.List;
+
+public interface HostRoomDBGateway {
+    List<RoomDBModel> getRooms();
+    void addARoom(RoomDBModel request);
+}
