+++ src/main/java/leave_room/LeaveRoomDBGateway.java
@@ -1,0 +1,8 @@
+package leave_room;
+
+import database.RoomDBModel;
+
+public interface LeaveRoomDBGateway {
+    void leaveARoom(RoomDBModel room);
+    RoomDBModel getARoom(Integer roomID);
+}
