+++ src/main/java/messaging/MessageDBGateway.java
@@ -1,0 +1,8 @@
+package messaging;
+
+import database.RoomDBModel;
+
+public interface MessageDBGateway {
+    RoomDBModel getARoom(Integer roomID);
+    void sendAMessage(RoomDBModel room);
+}
