+++ src/main/java/messaging/MessageInputBoundary.java
@@ -1,9 +1,6 @@
 package messaging;
 
-import models.MessageModel;
-
 public interface MessageInputBoundary {
     void handleSendMessage(MessageModel msgModel);
-
     void handleRetrieveMessages(MessageModel msgModel);
 }
