+++ src/main/java/signup_and_login/UserInteractor.java
@@ -1,8 +1,6 @@
 package signup_and_login;
-import database.UserDBGateway;
-import models.UserDBModel;
+import database.UserDBModel;
 import entities.User;
-import models.UserModel;
 
 import java.util.List;
