+++ src/test/java/signup_and_login/UserControllerTest.java
@@ -1,6 +1,5 @@
 package signup_and_login;
 
-import models.UserModel;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.*;
